Changed click behavior
Direct links are now plain links, stopping the click event from bubbling up to google's handler.
Click the plus sign for the picture and the source info for the site. Frees up CTRL+Click for standard behavior (open in new Tab), as the links now open in the same tab by default.